Add netstandard2.0 to Oakton.AspNetCore
This change is included in a set of changes with the intent of adding netstandard2.0 back to Jasper in the newest versions. The big requirement is the use of IHostBuilder in CommandLineHostingExtensions.RunOaktonCommands (instead of IHostBuilder).

Remove preprocessor redirects for netstandard2.0 that attempt to use Microsoft.AspNetCore.Hosting instead of Microsoft.Extensions.Hosting. Microsoft.Extensions.Hosting works with netstandard2.0. One other part where this is removed is in RunCommand, where the old netstandard2.0 libraries were still using IWebHost, as well as writing some console output. This has been removed to sync it with the netcore3.1 and net5.0 versions.

Add netstandard2.0 framework target to Oakton.AspNetCore.

Add Microsoft.Extensions.Hosting conditionally to netstandard2.0 target framework type. The version used is identical to netcoreapp3.1, so these could possibly be merged in the future.